refactor: restructure SkillExperienceChart component and move to dedicated directory
